Float

+++ Trabajo_2/src/trabajo_2/Trabajo_2.java
@@ -5,9 +5,9 @@ import javax.swing.JOptionPane;
 public class Trabajo_2 {
     public static void main(String[] args) {
         //Panel Principal
-        int n1;
-        int n2;
-        int resultado;
+        float n1;
+        float n2;
+        float resultado;
         n1=Integer.parseInt(JOptionPane.showInputDialog("Ingrese el primer número"));
         n2=Integer.parseInt(JOptionPane.showInputDialog("Ingrese el segundo número")); 
         int op=Integer.parseInt(JOptionPane.showInputDialog("Ingrese el segundo número\n1.- Suma\n2.- Resta\n3.- Multiplicación\n4.- División"));
